[Decomp][Spell] Cohesion split: extract cast pipeline to SpellCast.cpp
Move the Spell:: cast pipeline (PreCastCheck, SpellStart, Prepare, cancel,
cast, TakeAmmo, handle_immediate, handle_delayed, _handle_immediate_phase,
_handle_finish_phase, SendSpellCooldown, update, finish) out of Spell.cpp into
a new SpellCast.cpp TU. Contiguous block, bodies only; declarations stay in
Spell.h, so call sites and linkage are unchanged. No file-statics, no
file-local helper-class refs, no cross-TU template instantiations in range.
Spell.cpp 10806 -> 9676. Pure relocation, no behaviour change.
